Added Capp status sync addon.
Added two components:
manger -  responsible on the life cycle of agents on the managed cluster.
agent -responsible to update the Capp statuses on the hub cluster. The agent runs on the managed cluster.
There is also deployments section which holds the manifests for the deployment and the deploy it self is in the make file.

Signed-off-by: Ran Faingersh <fran2810@gmail.com>